# Conflicts:
#	src/main/java/com/sericulture/marketandauction/model/api/marketauction/FarmerReadyPaymentInfoForSeedMarketResponse.java
#	src/main/java/com/sericulture/marketandauction/service/FarmerPaymentService.java

+++ src/main/java/com/sericulture/marketandauction/model/entity/LotGroupage.java
@@ -37,7 +37,7 @@ public class LotGroupage extends BaseEntity implements Serializable {
     private Long amount;
 
     @Column(name = "MARKET_FEE")
-    private Long marketFee;
+    private Double marketFee;
 
     @Column(name = "SOLD_AMOUNT")
     private Long soldAmount;
